+++ ArkEcosystem.Client.Tests/API/TransactionsTest.cs
@@ -23,13 +23,13 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using ArkEcosystem.Client.API.Two;
-using ArkEcosystem.Client.API.Two.Models;
+using ArkEcosystem.Client.API;
+using ArkEcosystem.Client.API.Models;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace ArkEcosystem.Client.Tests.API.Two
+namespace ArkEcosystem.Client.Tests.API
 {
-    using Two = ArkEcosystem.Client.API.Two.Two;
+    using Two = ArkEcosystem.Client.API.Two;
 
     [TestClass]
     public class TransactionsTest
