+++ ArkEcosystem.Client/API/Blocks.cs
@@ -24,10 +24,10 @@ using System.Collections.Generic;
 using System.Net.Http;
 using System.Threading.Tasks;
 using Newtonsoft.Json.Linq;
-using ArkEcosystem.Client.API.Two.Models;
+using ArkEcosystem.Client.API.Models;
 using ArkEcosystem.Client.Helpers;
 
-namespace ArkEcosystem.Client.API.Two
+namespace ArkEcosystem.Client.API
 {
     public class Blocks
     {
